Isolating some random functions in a utils package
